==> code/1. Big Data Preprocessing.ipynb <==
# Import libraries

import pandas as pd
from feature_engine.selection import RecursiveFeatureElimination
from feature_engine.selection import DropCorrelatedFeatures
from feature_engine.selection import DropConstantFeatures

# Preprocess Water Index

mndwi_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Air/MNDWI/MNDWI_sumsel.csv', sep = ';')
mndwi_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Air/MNDWI/MNDWI_kalsel.csv', sep = ';')
mndwi = pd.concat([mndwi_sumsel, mndwi_kalsel], axis = 0)

ndwi_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Air/NDWI/NDWI_sumsel.csv', sep = ';')
ndwi_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Air/NDWI/NDWI_kalsel.csv', sep = ';')
ndwi = pd.concat([ndwi_sumsel, ndwi_kalsel], axis = 0)

# Preprocess Built Index

bui_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Keterbangunan/BUI/BUI_sumsel.csv', sep = ';')
bui_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Keterbangunan/BUI/BUI_kalsel.csv', sep = ';')
bui = pd.concat([bui_sumsel, bui_kalsel], axis = 0)

ndbi_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Keterbangunan/NDBI/NDBI_sumsel.csv', sep = ';')
ndbi_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Keterbangunan/NDBI/NDBI_kalsel.csv', sep = ';')
ndbi = pd.concat([ndbi_sumsel, ndbi_kalsel], axis = 0)

ndti_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Keterbangunan/NDTI/NDTI_sumsel.csv', sep = ';')
ndti_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Keterbangunan/NDTI/NDTI_kalsel.csv', sep = ';')
ndti = pd.concat([ndti_sumsel, ndti_kalsel], axis = 0)

ui_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Keterbangunan/UI/UI_sumsel.csv', sep = ';')
ui_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Keterbangunan/UI/UI_kalsel.csv', sep = ';')
ui = pd.concat([ui_sumsel, ui_kalsel], axis = 0)

vibi_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Keterbangunan/VIBI/VIBI_sumsel.csv', sep = ';')
vibi_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Keterbangunan/VIBI/VIBI_kalsel.csv', sep = ';')
vibi = pd.concat([vibi_sumsel, vibi_kalsel], axis = 0)

# Preprocess Vegetation Index

arvi_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/ARVI/ARVI_sumsel.csv', sep = ';')
arvi_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/ARVI/ARVI_kalsel.csv', sep = ';')
arvi = pd.concat([arvi_sumsel, arvi_kalsel], axis = 0)

evi_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/EVI/EVI_sumsel.csv', sep = ';')
evi_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/EVI/EVI_kalsel.csv', sep = ';')
evi = pd.concat([evi_sumsel, evi_kalsel], axis = 0)

gndvi_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/GNDVI/GNDVI_sumsel.csv', sep = ';')
gndvi_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/GNDVI/GNDVI_kalsel.csv', sep = ';')
gndvi = pd.concat([gndvi_sumsel, gndvi_kalsel], axis = 0)

lai_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/LAI/LAI_sumsel.csv', sep = ';')
lai_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/LAI/LAI_kalsel.csv', sep = ';')
lai = pd.concat([lai_sumsel, lai_kalsel], axis = 0)

ndvi_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/NDVI/NDVI_sumsel.csv', sep = ';')
ndvi_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/NDVI/NDVI_kalsel.csv', sep = ';')
ndvi = pd.concat([ndvi_sumsel, ndvi_kalsel], axis = 0)

savi_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/SAVI/SAVI_sumsel.csv', sep = ';')
savi_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/SAVI/SAVI_kalsel.csv', sep = ';')
savi = pd.concat([savi_sumsel, savi_kalsel], axis = 0)

sipi_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/SIPI/SIPI_sumsel.csv', sep = ';')
sipi_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Indeks Vegetasi/SIPI/SIPI_kalsel.csv', sep = ';')
sipi = pd.concat([sipi_sumsel, sipi_kalsel], axis = 0)

# Preprocess Infrastructure

jml_jln_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Infrastruktur/Jumlah Jalan/jumlah_jalan_sumsel.csv', sep = ';')
jml_jln_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Infrastruktur/Jumlah Jalan/jumlah_jalan_kalsel.csv', sep = ';')
jml_jln = pd.concat([jml_jln_sumsel, jml_jln_kalsel], axis = 0)

pjg_jln_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Infrastruktur/Panjang Jalan/panjang_jalan_sumsel.csv', sep = ';')
pjg_jln_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Infrastruktur/Panjang Jalan/panjang_jalan_kalsel.csv', sep = ';')
pjg_jln = pd.concat([pjg_jln_sumsel, pjg_jln_kalsel], axis = 0)

# Preprocess Nightlight Radiance

ntl_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Nightlight Radiance/NTL_sumsel.csv', sep = ';')
ntl_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Nightlight Radiance/NTL_kalsel.csv', sep = ';')
ntl = pd.concat([ntl_sumsel, ntl_kalsel], axis = 0)

# Preprocess Polution

co_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Polusi/CO/CO_sumsel.csv', sep = ';')
co_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Polusi/CO/CO_kalsel.csv', sep = ';')
co = pd.concat([co_sumsel, co_kalsel], axis = 0)

h2o_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Polusi/H2O/H2O_sumsel.csv', sep = ';')
h2o_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Polusi/H2O/H2O_kalsel.csv', sep = ';')
h2o = pd.concat([h2o_sumsel, h2o_kalsel], axis = 0)

no2_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Polusi/NO2/NO2_sumsel.csv', sep = ';')
no2_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Polusi/NO2/NO2_kalsel.csv', sep = ';')
no2 = pd.concat([no2_sumsel, no2_kalsel], axis = 0)

so2_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Polusi/SO2/SO2_sumsel.csv', sep = ';')
so2_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Polusi/SO2/SO2_kalsel.csv', sep = ';')
so2 = pd.concat([so2_sumsel, so2_kalsel], axis = 0)

# Preprocess RWI

rwi_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/RWI/rwi_sumsel.csv', sep = ';')
rwi_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/RWI/rwi_kalsel.csv', sep = ';')
rwi = pd.concat([rwi_sumsel, rwi_kalsel], axis = 0)

# Preprocess Climate

lst_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Temperatur Udara/LST/LST_sumsel.csv', sep = ';')
lst_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Temperatur Udara/LST/LST_kalsel.csv', sep = ';')
lst = pd.concat([lst_sumsel, lst_kalsel], axis = 0)

prec_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Temperatur Udara/PREC/PREC_sumsel.csv', sep = ';')
prec_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Temperatur Udara/PREC/PREC_kalsel.csv', sep = ';')
prec = pd.concat([prec_sumsel, prec_kalsel], axis = 0)

# Preprocess Land Cover

sum_esa_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Tutupan Lahan/jumlah_tutupanlahan_by_esa_sumsel.csv', sep = ';')
sum_esa_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Tutupan Lahan/jumlah_tutupanlahan_by_esa_kalsel.csv', sep = ';')
sum_esa = pd.concat([sum_esa_sumsel, sum_esa_kalsel], axis = 0)

perc_esa_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Tutupan Lahan/persen_tutupanlahan_by_esa_sumsel.csv', sep = ';')
perc_esa_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Tutupan Lahan/persen_tutupanlahan_by_esa_kalsel.csv', sep = ';')
perc_esa = pd.concat([perc_esa_sumsel, perc_esa_kalsel], axis = 0)

sum_esri_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Tutupan Lahan/jumlah_tutupanlahan_by_esri_sumsel.csv', sep = ';')
sum_esri_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Tutupan Lahan/jumlah_tutupanlahan_by_esri_kalsel.csv', sep = ';')
sum_esri = pd.concat([sum_esri_sumsel, sum_esri_kalsel], axis = 0)

perc_esri_sumsel = pd.read_csv('../dataset/auxiliary/Big Data/Tutupan Lahan/persen_tutupanlahan_by_esri_sumsel.csv', sep = ';')
perc_esri_kalsel = pd.read_csv('../dataset/auxiliary/Big Data/Tutupan Lahan/persen_tutupanlahan_by_esri_kalsel.csv', sep = ';')
perc_esri = pd.concat([perc_esri_sumsel, perc_esri_kalsel], axis = 0)

# Merge all file

aux_big_data = mndwi.merge(ndwi, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(bui, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(ndbi, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(ndti, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(ui, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(vibi, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(arvi, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(evi, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(gndvi, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(lai, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(ndvi, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(savi, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(sipi, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(jml_jln, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(pjg_jln, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(ntl, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(co, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(h2o, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(no2, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(so2, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(rwi, left_on='idkec', right_on='geo_id')
aux_big_data = aux_big_data.merge(lst, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(prec, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(sum_esa, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(perc_esa, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(sum_esri, left_on='idkec', right_on='idkec')
aux_big_data = aux_big_data.merge(perc_esri, left_on='idkec', right_on='idkec')

aux_big_data = aux_big_data.drop(['nmkec_x', 'nmkec_y'], axis = 1)
aux_big_data = aux_big_data.rename(columns={"idkec": "Kecamatan"})
aux_big_data.to_csv('../dataset/auxiliary/Big Data/auxiliary.csv', index = False)
