==> code/2. Mixed Data Preprocessing.ipynb <==
# Import libraries

import pandas as pd

# Import Podes data and Big Data

podes = pd.read_excel('../dataset/auxiliary/Pendataan PODES 2021/data_podes 2021_1663.xlsx')
bigdata = pd.read_csv('../dataset/auxiliary/Big Data/auxiliary.csv')

auxiliary = podes.merge(bigdata, left_on = 'Kecamatan', right_on = 'Kecamatan', how = 'left')
auxiliary.to_csv('../dataset/auxiliary/auxiliary.csv', index = False)
