==> code/3. Linear Variable Selection for Podes.ipynb <==
# Variable Selection for Global Linear Regression

import pandas as pd
from sklearn.linear_model import LinearRegression
from feature_engine.selection import RecursiveFeatureElimination
from feature_engine.selection import SmartCorrelatedSelection
from feature_engine.selection import DropConstantFeatures
from sklearn.preprocessing import StandardScaler

indikator = pd.read_csv('../dataset/indikator/DE_INDIKATOR3.csv', sep = ';')
auxiliary = pd.read_excel('../dataset/auxiliary/Pendataan PODES 2021/data_podes 2021_1663.xlsx')

kecamatan = auxiliary['Kecamatan']
auxiliary = auxiliary.dropna(axis = 1)

column = auxiliary.columns
scaler = StandardScaler()
auxiliary = scaler.fit_transform(auxiliary)
auxiliary = pd.DataFrame(auxiliary, columns = column)

auxiliary = auxiliary.drop(['Kecamatan'], axis = 1)
auxiliary = pd.concat([kecamatan, auxiliary], axis = 1)

dataset = indikator.merge(auxiliary, left_on = 'Kecamatan', right_on = 'Kecamatan', how = 'left')
dataset_1 = dataset.drop(['Unnamed: 0', 'Provinsi', 'Kecamatan', 'SE', 'VAR', 'CI LOWER', 'CI UPPER', 'RSE', 'DEFF', 'r101', 'r102', 'r103'], axis = 1)

y = dataset_1['Rata2 Kapita']
X = dataset_1.drop(['Rata2 Kapita'], axis = 1)

tr = DropConstantFeatures(tol=0.7)
Xt = tr.fit_transform(X)

tr = SmartCorrelatedSelection(
    method="pearson",
    threshold=0.8,
    selection_method="variance",
)
Xt = tr.fit_transform(Xt)

regressor = LinearRegression()
tr = RecursiveFeatureElimination(estimator = regressor, scoring='neg_root_mean_squared_error', cv = 5)
Xt = tr.fit_transform(Xt, y)
Xt.to_csv('../dataset/auxiliary/Pendataan PODES 2021/auxiliary_linear_transformed.csv', index = False)

# Variable Selection for Sumsel Linear Regression

indikator = pd.read_csv('../dataset/indikator/DE_INDIKATOR3.csv', sep = ';')
auxiliary = pd.read_excel('../dataset/auxiliary/Pendataan PODES 2021/data_podes 2021_1663.xlsx')

kecamatan = auxiliary['Kecamatan']
auxiliary = auxiliary.dropna(axis = 1)

column = auxiliary.columns
scaler = StandardScaler()
auxiliary = scaler.fit_transform(auxiliary)
auxiliary = pd.DataFrame(auxiliary, columns = column)

auxiliary = auxiliary.drop(['Kecamatan'], axis = 1)
auxiliary = pd.concat([kecamatan, auxiliary], axis = 1)

dataset = indikator.merge(auxiliary, left_on = 'Kecamatan', right_on = 'Kecamatan', how = 'left')
dataset = dataset[dataset['Provinsi'] == 16] 
dataset_1 = dataset.drop(['Unnamed: 0', 'Provinsi', 'Kecamatan', 'SE', 'VAR', 'CI LOWER', 'CI UPPER', 'RSE', 'DEFF', 'r101', 'r102', 'r103'], axis = 1)

y = dataset_1['Rata2 Kapita']
X = dataset_1.drop(['Rata2 Kapita'], axis = 1)

tr = DropConstantFeatures(tol=0.7)
Xt = tr.fit_transform(X)

tr = SmartCorrelatedSelection(
    method="pearson",
    threshold=0.8,
    selection_method="variance",
)
Xt = tr.fit_transform(Xt)

regressor = LinearRegression()
tr = RecursiveFeatureElimination(estimator = regressor, scoring='neg_root_mean_squared_error', cv = 5)
Xt = tr.fit_transform(Xt, y)
Xt.to_csv('../dataset/auxiliary/Pendataan PODES 2021/auxiliary_sumsel_linear_transformed.csv', index = False)

# Variable Selection for Kalsel Linear Regression

indikator = pd.read_csv('../dataset/indikator/DE_INDIKATOR3.csv', sep = ';')
auxiliary = pd.read_excel('../dataset/auxiliary/Pendataan PODES 2021/data_podes 2021_1663.xlsx')

kecamatan = auxiliary['Kecamatan']
auxiliary = auxiliary.dropna(axis = 1)

column = auxiliary.columns
scaler = StandardScaler()
auxiliary = scaler.fit_transform(auxiliary)
auxiliary = pd.DataFrame(auxiliary, columns = column)

auxiliary = auxiliary.drop(['Kecamatan'], axis = 1)
auxiliary = pd.concat([kecamatan, auxiliary], axis = 1)

dataset = indikator.merge(auxiliary, left_on = 'Kecamatan', right_on = 'Kecamatan', how = 'left')
dataset = dataset[dataset['Provinsi'] == 63] 
dataset_1 = dataset.drop(['Unnamed: 0', 'Provinsi', 'Kecamatan', 'SE', 'VAR', 'CI LOWER', 'CI UPPER', 'RSE', 'DEFF', 'r101', 'r102', 'r103'], axis = 1)

y = dataset_1['Rata2 Kapita']
X = dataset_1.drop(['Rata2 Kapita'], axis = 1)

tr = DropConstantFeatures(tol=0.7)
Xt = tr.fit_transform(X)

tr = SmartCorrelatedSelection(
    method="pearson",
    threshold=0.8,
    selection_method="variance",
)
Xt = tr.fit_transform(Xt)

regressor = LinearRegression()
tr = RecursiveFeatureElimination(estimator = regressor, scoring='neg_root_mean_squared_error', cv = 5)
Xt = tr.fit_transform(Xt, y)
Xt.to_csv('../dataset/auxiliary/Pendataan PODES 2021/auxiliary_kalsel_linear_transformed.csv', index = False)
